requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
django==4.1.3
django-ninja==0.19.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pydantic==1.10.2 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
typing-extensions==4.4.0 ; python_version >= '3.7'
